li tawa e ilo pi toki e nimi wawa tawa jan lon poki ante toki.hpp(en .cpp).
